fix: Restyle event page & fix events detail
