+++ _events/20230404.mdx
@@ -4,20 +4,22 @@ place: The Space, Via dei Lucani 1 Latina
 maps: https://maps.app.goo.gl/4RXKYzYeYPTUkeXBA
 thumbnail: /assets/events/20230404.png
 title: Introduzione Alla Community & Lightning Talks
-description: La community di informatici Pontini, nata nei primi giorni del 2023, finalmente esce allo scoperto, con il primo evento ufficiale! Dopo un'introduzione Alla Community, seguiranno dei Lightning talks dei membri della community!
-prerequisites:
-  [
-    'Node.js installed on your computer',
-    'Basic knowledge working with Next.js and TypeScript',
-  ]
-stacks: ['Next.js', 'TypeScript', 'Git']
+description: "🏁 La community di informatici Pontini, nata nei primi giorni del 2023, finalmente esce allo scoperto, con il primo evento ufficiale!<br />Dopo un'introduzione Alla Community, seguiranno dei Lightning talks dei membri della community!"
+tags: ['Next.js', 'TypeScript', 'Git']
 ---
 
-<Prerequisites />
-<Stacks />
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+Arriva il primo evento della community nata i primi giorni del 2023, Latina In Tech!
+
+
+Un evento all'insegna della condivisione, dove verrà introdotta la community e l'idea che c'è dietro da parte del fondatore, **[Andrea Coluzzi](https://www.coluzziandrea.com/)**
+
+Dopodiché ascolteremo le voci degli speaker appartenenti alla community stessa:
+
+
+ \- **[Matteo Boschi](https://www.linkedin.com/in/matteo-boschi/)** 
+ 
+ \- **[Luca Visentin](https://www.linkedin.com/in/visentin-luca/)** 
+
+ \- **[Gabriele Quaresima](https://www.linkedin.com/in/gabriele-quaresima/)** 
+
+ \- **[Lorenzo Ceglia](https://www.linkedin.com/in/lorenzo-ceglia/)**
