+++ _events/20230509.mdx
@@ -4,20 +4,22 @@ place: The Space, Via dei Lucani 1 Latina
 maps: https://maps.app.goo.gl/4RXKYzYeYPTUkeXBA
 thumbnail: /assets/events/20230509.png
 title: Lightning Talks Vol. 2
-description: La community LiT torna con il secondo appuntamento della stagione 2023, con nuovi speaker! Il programma della serata è il seguente - Antonio Fabrizio con Startup & Imprenditorialità, Fabrizio Cafolla con Open source e progetti, Alessandro Marino con Journey to Cloud Computing e Matteo Benin con Flutter
-prerequisites:
-  [
-    'Node.js installed on your computer',
-    'Basic knowledge working with Next.js and TypeScript',
-  ]
-stacks: ['Next.js', 'TypeScript', 'Git']
+description: "❗ La community LiT torna con il secondo appuntamento della stagione 2023, con nuovi speaker!<br />Per questa serata avremo come speaker: **[Antonio Fabrizio](https://www.linkedin.com/in/fabrizioantonio/)**, **[Fabrizio Cafolla](https://www.linkedin.com/in/fabrizio-cafolla/)**, **[Alessandro Marino](https://www.linkedin.com/in/alessandro-marino-ac/)**, **[Anna Maria Coppola](https://www.linkedin.com/in/annamaria-coppola-9a301613/)** e **[Matteo Benin](https://www.linkedin.com/in/matteo-benin-28b11195/)**"
+tags: ['Next.js', 'TypeScript', 'Git']
 ---
 
-<Prerequisites />
-<Stacks />
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+Arriva il secondo appuntamento con il format **Lightning Talks** della community Latina In Tech.
+
+A farci da cornice sempre il nostro amato coworking **[The Space](https://www.thespacecoworking.website/)**
+
+Questa volta gli speaker saranno cinque:
+
+ \- Startup & Imprenditorialità con **[Antonio Fabrizio](https://www.linkedin.com/in/fabrizioantonio/)** collegato da remoto
+
+ \- Open source e progetti con **[Fabrizio Cafolla](https://www.linkedin.com/in/fabrizio-cafolla/)**
+ 
+ \- Journey to Cloud Computing con **[Alessandro Marino](https://www.linkedin.com/in/alessandro-marino-ac/)**
+
+ \- Coding Dojo e TDD con **[Anna Maria Coppola](https://www.linkedin.com/in/annamaria-coppola-9a301613/)**
+
+ \- Flutter con **[Matteo Benin](https://www.linkedin.com/in/matteo-benin-28b11195/)**
