+++ _events/20230606.mdx
@@ -5,20 +5,18 @@ maps: https://maps.app.goo.gl/4RXKYzYeYPTUkeXBA
 youtubeUrl: https://www.youtube.com/watch?v=araCZjntJRw
 thumbnail: /assets/events/20230606.png
 title: Lightning Talks Vol. 3
-description:  Siamo entusiasti di annunciarvi il prossimo Lightning Talks della community LiT! Si parlerà di Cost Optimization on Google Cloud con Sergio Picca, Panoramica delle Reti mobili e 5G con Alessio Marinelli e KubeCon & operatori Kubernetes con Gabriele Quaresima
-prerequisites:
-  [
-    'Node.js installed on your computer',
-    'Basic knowledge working with Next.js and TypeScript',
-  ]
-stacks: ['Next.js', 'TypeScript', 'Git']
+description:  "📢 Siamo entusiasti di annunciarvi il prossimo Lightning Talks della community LiT!<br />Si parlerà di Cost Optimization on Google Cloud con **[Sergio Picca](https://www.linkedin.com/in/sergio-picca/)**, Panoramica delle Reti mobili e 5G con **[Alessio Marinelli](https://www.linkedin.com/in/alessio-marinelli/)** e KubeCon & operatori Kubernetes con **[Gabriele Quaresima](https://www.linkedin.com/in/gabriele-quaresima/)**"
+tags: ['Next.js', 'TypeScript', 'Git']
 ---
 
-<Prerequisites />
-<Stacks />
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+Arriva il terzo appuntamento con il format Lightning Talks della community Latina In Tech.
+
+A farci da cornice sempre il nostro amato coworking **[The Space](https://www.thespacecoworking.website/)**
+
+Questa volta gli speaker saranno tre:
+
+ \- Cost Optimization on Google Cloud con **[Sergio Picca](https://www.linkedin.com/in/sergio-picca/)**
+
+ \- Panoramica delle Reti mobili e 5G con **[Alessio Marinelli](https://www.linkedin.com/in/alessio-marinelli/)** 
+ 
+ \- KubeCon & operatori Kubernetes con **[Gabriele Quaresima](https://www.linkedin.com/in/gabriele-quaresima/)**
