+++ _events/20230629.mdx
@@ -4,20 +4,8 @@ place: Viridex s.r.l., Via Gennaro Del Prete 2/4 Cisterna di Latina
 maps: https://maps.app.goo.gl/j3t4TmCTohgrCSbF6
 thumbnail: /assets/events/20230629.png
 title: La Giusta Relazione di Lavoro
-description: La community di informatici Pontini torna per un ultimo evento prima dell'estate, questa volta con un ospite d'onore! Direttamente da Treviso, avremo l'opportunità di ascoltare Antonio Dell'Ava, Senior Frontend Dev in eDreams ODIGEO! Antonio ci parlerà della 'Giusta Relazione di Lavoro - Cosa è, Perché succede, Quando cambiare, Come fare a crearla'
-prerequisites:
-  [
-    'Node.js installed on your computer',
-    'Basic knowledge working with Next.js and TypeScript',
-  ]
-stacks: ['Next.js', 'TypeScript', 'Git']
+description: "🏖️ La community di informatici Pontini torna per un ultimo evento prima dell'estate, questa volta con un ospite d'onore!<br />🚄 Direttamente da Treviso, avremo l'opportunità di ascoltare **[Antonio Dell'Ava](https://www.linkedin.com/in/antoniodellava/)**, Senior Frontend Dev in eDreams ODIGEO!"
+tags: ['Next.js', 'TypeScript', 'Git']
 ---
 
-<Prerequisites />
-<Stacks />
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+Antonio ci parlerà di: **La Giusta Relazione di Lavoro - Cosa è, Perché succede, Quando cambiare, Come fare a crearla**
