+++ _events/20230922.mdx
@@ -4,20 +4,12 @@ place: The Space, Via dei Lucani 1 Latina
 maps: https://maps.app.goo.gl/4RXKYzYeYPTUkeXBA
 thumbnail: /assets/events/20230922.webp
 title: Under the hood at Klarna - Managing digital gift cards and leveraging AI
-description: La stagione 2023/2024 della community riparte alla grande, con una collaborazione incredibile! Infatti per l'evento di settembre siamo riusciti ad ottenere una sponsorizzazione da Klarna Engineering. Non è finita qui, perchè avremo come super ospite un Lead Engineer dell'azienda, che ci raggiungerà direttamente da Milano - Antonio Trapani! Antonio ci parlerà di Intelligenza Artificiale e come viene applicata nei progetti Klarna. Inoltre anche Andrea Coluzzi parlerà del suo team, il team Gift Cards!
-prerequisites:
-  [
-    'Node.js installed on your computer',
-    'Basic knowledge working with Next.js and TypeScript',
-  ]
-stacks: ['Next.js', 'TypeScript', 'Git']
+description: "⚡️ La stagione 2023/2024 della community riparte alla grande, con una collaborazione incredibile!<br />Per l'evento di settembre infatti siamo riusciti ad ottenere una sponsorizzazione da **[Klarna Engineering](https://engineering.klarna.com/)**"
+tags: ['Next.js', 'TypeScript', 'Git']
 ---
 
-<Prerequisites />
-<Stacks />
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+Non è finita qui, perchè avremo come super ospite un **Lead Engineer** dell'azienda, che ci raggiungerà direttamente da Milano - **[Antonio Trapani](https://www.linkedin.com/in/antoniotrapani/)**! 
+<br />
+Antonio ci parlerà di Intelligenza Artificiale e come viene applicata nei progetti Klarna. 
+<br />
+Inoltre anche **[Andrea Coluzzi](https://www.linkedin.com/in/andrea-coluzzi)** parlerà del suo team in Klarna, il team Gift Cards!
