+++ _events/20231129.mdx
@@ -6,13 +6,8 @@ maps: https://maps.app.goo.gl/UMP3jFwA5rihVsEj7
 youtubeUrl: https://www.youtube.com/watch?v=81DfVfQNycI
 thumbnail: /assets/events/20231129.jpg
 title: Presentazione Startup Lazio Innova
-description: "Diversi ospiti per un evento di connessione con imprenditoria locale.<br/><br/>**[Francesco Paolo Russo @ToBe s.r.l.](https://tobe-srl.it/)** ci presenterà soluzioni LiFi, tecnologia di trasmissione dati wireless attraverso la modulazione della luce LED.<br/><br/>**[Domenico Crescenzo @Screevo](https://screevo.ai/)** ci mostrerà come l'intelligenza artificiale può essere impiegata negli algoritmo per la tecnologia vocale.<br/><br/>Parteciperanno inoltre **Maurizio Andolfi** responsabile dello Spazio Attivo di Latina e **Martina Becchimanzi**  responsabile e coordinatrice di OpenHub Latina.<br/><br/>Dopo le presentazioni, avremo un momento di networking in cui potrai incontrare e interagire con i fondatori delle startup, gli ospiti d'onore e altri partecipanti interessati."
-prerequisites:
-  [
-    'Node.js installed on your computer',
-    'Basic knowledge working with Next.js and TypeScript',
-  ]
-stacks: ['Next.js', 'TypeScript', 'Git']
+description: "🎯 Per questo evento avremo diversi ospiti rappresentanti di alcune startup pontine per un evento di connessione con imprenditoria locale targato **[Lazio Innova](https://www.lazioinnova.it/)**"
+tags: ['Next.js', 'TypeScript', 'Git']
 speakers: 
   [
     '{"name": "Francesco Paolo Russo", "role": "CEO", "company": "ToBe s.r.l.", "thumbnail": "/assets/speakers/francesco-paolo-russo.jpeg", "linkedinUrl": "https://it.linkedin.com/in/francescopaolorusso?trk=public_post_feed-actor-name"}',
@@ -20,11 +15,4 @@ speakers:
   ]
 ---
 
-<Prerequisites />
-<Stacks />
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+**[Francesco Paolo Russo @ToBe s.r.l.](https://tobe-srl.it/)** ci presenterà soluzioni LiFi, tecnologia di trasmissione dati wireless attraverso la modulazione della luce LED.<br/><br/>**[Domenico Crescenzo @Screevo](https://screevo.ai/)** ci mostrerà come l'intelligenza artificiale può essere impiegata negli algoritmo per la tecnologia vocale.<br/><br/>Parteciperanno inoltre **Maurizio Andolfi** responsabile dello Spazio Attivo di Latina e **Martina Becchimanzi**  responsabile e coordinatrice di OpenHub Latina.<br/><br/>Dopo le presentazioni, avremo un momento di networking in cui potrai incontrare e interagire con i fondatori delle startup, gli ospiti d'onore e altri partecipanti interessati.
