+++ _events/20231221.mdx
@@ -6,20 +6,12 @@ maps: https://maps.app.goo.gl/Gf2JkbTWHC4rk9Zj6
 youtubeUrl: https://www.youtube.com/watch?v=GNd-L3qPdPE
 thumbnail: /assets/events/20231221.png
 title: Serverless senza segreti con le Azure Functions, Node.js, GitHub Actions e Codespaces!
-description: "🌲 Siamo felici di presentarvi l'evento natalizio della community **Latina In Tech**!<br/>🍾 Festeggeremo l'anno passato proprio nella location dove è stato tenuto il primo evento LiT: il nostro coworking di fiducia The Space!!!<br/>⚡️ Per questa occasione avremo un ospite speciale, Autore, Tech Trainer & Speaker a livello internazionale: **[Massimo Bonanni](https://www.linkedin.com/in/massimobonanni/)**!<br/>🔥 Massimo ha una grandissima esperienza sul Cloud, essendo uno dei punti di riferimento per quanto riguarda Azure!."
-prerequisites:
+description: "🌲 Siamo felici di presentarvi l'evento natalizio della community!<br />🍾 Festeggeremo l'anno passato proprio nella location dove è stato tenuto il primo evento: il nostro coworking di fiducia **[The Space](https://www.thespacecoworking.website/)**<br/>⚡️ Per questa occasione avremo un ospite speciale, Autore, Tech Trainer & Speaker a livello internazionale!"
+tags: ['Azure', 'NodeJS', 'Github', 'Serverless']
+speakers:
   [
-    'Node.js installed on your computer',
-    'Basic knowledge working with Next.js and TypeScript',
+    '{"name": "Massimo Bonanni", "role": "Technical Trainer & Speaker", "company": "Microsoft", "thumbnail": "/assets/speakers/bonanni.jpeg", "linkedinUrl": "https://www.linkedin.com/in/massimobonanni/"}'
   ]
-stacks: ['Azure', 'NodeJS', 'Github', 'Serverless']
 ---
 
-<Prerequisites />
-<Stacks />
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+**[Massimo Bonanni](https://www.linkedin.com/in/massimobonanni/)** ha una grandissima esperienza sul Cloud, essendo uno dei punti di riferimento per quanto riguarda Azure!.
