+++ _events/20240116.mdx
@@ -3,18 +3,14 @@ date: '2024-01-16T18:00:00.000Z'
 duration: 90
 place: Viridex s.r.l., Via Gennaro Del Prete 2/4, Cisterna di Latina (LT)
 maps: https://maps.app.goo.gl/33ZBXtVVSaDp5sEd6
+signup: https://www.eventbrite.com/e/biglietti-software-for-millions-of-users-783814238487
 thumbnail: /assets/events/20240116.jpg
 title: Software for millions of users!
-description: "🌱 Sei pronto a dare il benvenuto al 2024? noi si, e stavolta non puoi mancare!<br/><br/>Ad inaugurare il nuovo anno ci aiuteranno due engineers di [PagoPA](https://www.pagopa.it/it/), azienda che fornisce servizi a milioni di cittadini italiani, e forse anche a te!<br/><br/>[PagoPA](https://www.pagopa.it/it/), in particolare con la [piattaforma IO](https://ioapp.it/it), mette a disposizione un unico punto di accesso per interagire in modo semplice e sicuro con i servizi pubblici locali e nazionali, direttamente da mobile app.<br/><br/>⚡️ **[Emanuele De Cupis](https://www.linkedin.com/in/emanueledecupis/)** - _Staff Software Engineer_ ci spiegherà, come se parlasse a suo figlio, la programmazione funzionale, come può essere utile a chi sviluppa applicazioni web, che esperienza dovrebbe avere un team per approcciarla, ed altro ancora. Ce lo racconterà per esempi con fp-ts, la popolare libreria TypeScript dal cuore 🇮🇹.<br/><br/>⚡️ **[Alessio Dore](https://www.linkedin.com/in/alessio-dore/)** - _Head of IO Platform_ invece ci mostrerà come gestire ed utilizzare tecniche di rollout in applicazioni enterprise mono-ambiente: implementare miglioramenti graduati e progressivi senza dover passare attraverso fasi di verifica separate, come ad esempio staging o collaudo.<br/><br/>🏢 A fare da cornice al nostro evento, ringraziamo [Viridex](https://maps.app.goo.gl/33ZBXtVVSaDp5sEd6) che ci ospita di nuovo nei suoi uffici!"
+description: "🌱 Sei pronto a dare il benvenuto al 2024? noi si, e stavolta non puoi mancare!<br/>Ad inaugurare il nuovo anno ci aiuteranno due engineers di [PagoPA](https://www.pagopa.it/it/), azienda che fornisce servizi a milioni di cittadini italiani, e forse anche a te!"
 speakers:
   [
     '{"name": "Emanuele De Cupis", "role": "Staff Software Engineer", "company": "PagoPA S.p.A.", "thumbnail": "/assets/speakers/20240116/decupis.png", "linkedinUrl": "https://www.linkedin.com/in/emanueledecupis/"}',
     '{"name": "Alessio Dore", "role": "Head of IO Platform", "company": "PagoPA S.p.A.", "thumbnail": "/assets/speakers/20240116/dore.png", "linkedinUrl": "https://www.linkedin.com/in/alessio-dore/"}',
   ]
 ---
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
-incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
-nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
-Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
-fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
-culpa qui officia deserunt mollit anim id est laborum.
+[PagoPA](https://www.pagopa.it/it/), in particolare con la [piattaforma IO](https://ioapp.it/it), mette a disposizione un unico punto di accesso per interagire in modo semplice e sicuro con i servizi pubblici locali e nazionali, direttamente da mobile app.<br/><br/>⚡️ **[Emanuele De Cupis](https://www.linkedin.com/in/emanueledecupis/)** - _Staff Software Engineer_ ci spiegherà, come se parlasse a suo figlio, la programmazione funzionale, come può essere utile a chi sviluppa applicazioni web, che esperienza dovrebbe avere un team per approcciarla, ed altro ancora. Ce lo racconterà per esempi con fp-ts, la popolare libreria TypeScript dal cuore 🇮🇹.<br/><br/>⚡️ **[Alessio Dore](https://www.linkedin.com/in/alessio-dore/)** - _Head of IO Platform_ invece ci mostrerà come gestire ed utilizzare tecniche di rollout in applicazioni enterprise mono-ambiente: implementare miglioramenti graduati e progressivi senza dover passare attraverso fasi di verifica separate, come ad esempio staging o collaudo.<br/><br/>🏢 A fare da cornice al nostro evento, ringraziamo [Viridex](https://maps.app.goo.gl/33ZBXtVVSaDp5sEd6) che ci ospita di nuovo nei suoi uffici!
