+++ package.json
@@ -25,7 +25,6 @@
     "gray-matter": "^4.0.3",
     "luxon": "^3.4.4",
     "next": "14.0.4",
-    "next-mdx-remote": "^4.4.1",
     "npm-check-updates": "^16.14.11",
     "postcss": "^8.4.32",
     "react": "18.2.0",
