+++ src/pages/_app.tsx
@@ -1,6 +1,5 @@
 import 'tailwindcss/tailwind.css';
 import type { AppProps } from 'next/app';
-import { MdxComponentsProvider } from '@/context/mdxContext';
 import Head from 'next/head';
 import React from 'react';
 
@@ -27,9 +26,7 @@ export default function App({ Component, pageProps }: AppProps) {
         />
         <link rel='manifest' href='/site.webmanifest' />
       </Head>
-      <MdxComponentsProvider>
-        <Component {...pageProps} />
-      </MdxComponentsProvider>
+      <Component {...pageProps} />
     </>
   );
 }
